Introduce build-graph and read-graph commands
Sets up the project with two commands, build-graph and read-graph.
The latter is a placeholder for now.

build-graph reads in a fixed file of words of length 5.
It computes a graph of word indexes from the file, where the edges of the graph represent two words which have no letters in common.
The graph is serialized to an output file.