Remove duplicates from graph
Changes the data stored in the graph from word indexes to word bitmaps.
This removes a bunch of duplicated data from our graph, as we don't need individual nodes for two words which are anagrams.
E.g. the words `caged` and `cadge` will have the same edges, so why calculate their edges twice, store them twice, and traverse them twice?
We leverage a HashSet to deduplicate the bitmaps before generating the graph.