Introduce `build-solutions` and `read-solutions`
Reworks read-graph into build-solutions.
Leverages the graph to perform a depth-first traversal of the nodes to look for solutions.
Uses a hashset to keep track of nodes which are completely traversed, and stores solutions in btrees to deduplicate combinations.

This solver runs in about 3 hours on the current dataset.
It does come back with the correct 538 set of words (anagrams count as 1) that the video first found, including lots of `fldxt`.

Serialized solutions are read with `read-solutions`, which converts the bitsets back into the original string words.

Introduced some shared code into `lib.rs`.
Lots more can be pulled here, and sub modules created, with more tweaks.