Optimization by creating a DAG
Got stumped, and went looking for the better solution.
Couldn't find, so I rewatched the video.
The source description gives away the key to reducing tree traversal: only traverse in increasing node order.
That is, traverse the graph as a DAG.

We filter out neighbor nodes at graph construction, only including neighbors which are larger than the current node.
That reduces the number of ways to traverse the nodes abcde from 5! to 1.
This then cuts the runtime from hours to minutes, solving the problem in under 2 minutes.

Did almost commit a bug though.
Needed to remove the `completed` hashset.
This optimization was intended to reduce traversals when all combinations / paths were possible.
With the DAG and the randomness of where traversal could start, `completed` nodes were bisecting the graph, causing future traversals to miss 80% of potential solutions.